Better checks on file existence and creation of temporary folder

+++ src/Writer/ExcelWriter.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace FriendsOfSylius\SyliusImportExportPlugin\Writer;
 
+use FriendsOfSylius\SyliusImportExportPlugin\Exception\ExporterException;
 use FriendsOfSylius\SyliusImportExportPlugin\Exception\InvalidOrderException;
 use Port\Excel\ExcelWriter as PortExcelWriter;
 
@@ -59,6 +60,9 @@ class ExcelWriter implements WriterInterface
         $this->finish();
 
         $contents = file_get_contents($this->filename);
+        if (false === $contents) {
+            throw new ExporterException(sprintf('File %s could not be loaded', $this->filename));
+        }
 
         unlink($this->filename);
 
@@ -68,7 +72,11 @@ class ExcelWriter implements WriterInterface
     private function prepare(): void
     {
         if (null === $this->filename) {
-            $this->filename = tempnam($this->temporaryFolder, 'exp');
+            $tmp = tempnam($this->temporaryFolder, 'exp');
+            if (false === $tmp) {
+                throw new ExporterException(sprintf('Could not create temporary file in "%s"', $this->temporaryFolder));
+            }
+            $this->filename = $tmp;
         }
 
         if (null === $this->writer) {
